Fix bug causing feature test failures in Python3
type('')==bytes evaluates to True in Python2, and False in Python3. This
was causing some variable to never be set, which eventually led to
various silent errors downstream; the ultimate result was the
uncertainty values not being properly loaded from the data, which caused
all features which used those to have the wrong values.